[mtouch/mmp] Move some code to be shared.
