+++ tools/linker/MonoTouch.Tuner/MonoTouchSweepStep.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using Mono.Cecil;
 using Mono.Tuner;
 using Xamarin.Linker;
-using Xamarin.Tuner;
 
 namespace MonoTouch.Tuner {
 	
@@ -14,12 +13,6 @@ namespace MonoTouch.Tuner {
 		public MonoTouchSweepStep (LinkerOptions options)
 			: base (options.LinkSymbols)
 		{
-		}
-
-		protected DerivedLinkContext LinkContext {
-			get {
-				return (DerivedLinkContext) base.Context;
-			}
 		}
 
 		protected override void SweepAssembly (AssemblyDefinition assembly)
@@ -56,16 +49,6 @@ namespace MonoTouch.Tuner {
 				if (!need_ivt)
 					attributes.RemoveAt (i--);
 			}
-		}
-
-		protected override void InterfaceRemoved (TypeDefinition type, InterfaceImplementation iface)
-		{
-			base.InterfaceRemoved (type, iface);
-
-			// The static registrar needs access to the interfaces for protocols, so keep them around.
-			if (!LinkContext.ProtocolImplementations.TryGetValue (type, out var list))
-				LinkContext.ProtocolImplementations [type] = list = new List<TypeDefinition> ();
-			list.Add (iface.InterfaceType.Resolve ());
 		}
 	}
 }
